% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CL_gen_invbijection_table}
\alias{CL_gen_invbijection_table}
\title{Generate Inverse Bijection Table}
\usage{
CL_gen_invbijection_table(K, M, nClass)
}
\arguments{
\item{K}{Number of attributes.}

\item{M}{Level of attribute.}

\item{nClass}{Number of classes.}
}
\value{
A matrix representing the generated inverse bijection table.
}
\description{
Generate a table of inverse bijection vectors for given parameters.

This function generates a table of inverse bijection vectors for a specified number of attributes (K), level of attribute (M), and number of classes (nClass).
}
