% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeLB}
\alias{computeLB}
\title{Compute Lower Bound}
\usage{
computeLB(nClass, LBtable, Atable, p, betaj, Bp, Bindices)
}
\arguments{
\item{nClass}{Number of classes.}

\item{LBtable}{Matrix representing the lower bound constraints.}

\item{Atable}{Matrix representing the attribute profile table.}

\item{p}{Index of the parameter.}

\item{betaj}{Row vector of item response coefficient matrix B.}

\item{Bp}{Bp parameter.}

\item{Bindices}{Vector of latent class indices.}
}
\value{
The computed lower bound.
}
\description{
Compute the lower bound for a given parameter.

This function computes the lower bound for a given parameter based on provided tables and coefficients.
}
