% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computePYa}
\alias{computePYa}
\title{Compute Conditional Probabilities}
\usage{
computePYa(J, P, nClass, Atable, BETA, Kappa)
}
\arguments{
\item{J}{Number of items.}

\item{P}{Number of levels for each item.}

\item{nClass}{Number of classes.}

\item{Atable}{Matrix representing the attribute profile table.}

\item{BETA}{Matrix representing the item response coefficient matrix.}

\item{Kappa}{Matrix representing the item response threshold matrix.}
}
\value{
A list containing the adjusted item response coefficient matrix (ABETA) and the cube of conditional probabilities (PY_a).
}
\description{
Compute conditional probabilities based on provided parameters.

This function computes conditional probabilities based on the provided attribute profile table, item response coefficient matrix, and item response threshold matrix.
}
