% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{djpsamp}
\alias{djpsamp}
\title{Update Delta}
\usage{
djpsamp(Bplb, omega, sigma_B, Bp, ApApp, ApAb, ApZjp, term_h)
}
\arguments{
\item{Bplb}{Lower bound for Bp.}

\item{omega}{Omega parameter.}

\item{sigma_B}{Standard deviation for B.}

\item{Bp}{Bp parameter.}

\item{ApApp}{ApApp parameter.}

\item{ApAb}{ApAb parameter.}

\item{ApZjp}{ApZjp parameter.}

\item{term_h}{Additional term related to H.}
}
\value{
A double value.
}
\description{
Sample delta.

This function samples from the DJP (Double Joint Probability) distribution.
}
