% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_bijectionvector}
\alias{gen_bijectionvector}
\title{Get Bijection Vector}
\usage{
gen_bijectionvector(K, M)
}
\arguments{
\item{K}{Number of attributes.}

\item{M}{Level of attribute.}
}
\value{
The bijection vector.
}
\description{
Get the bijection vector for a given number of attributes and level of attribute.

This function calculates the bijection vector for a specified number of attributes (K) and level of attribute (M).
}
\examples{
gen_bijectionvector(3, 2)
}
