% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_Y}
\alias{gen_Y}
\title{Generate Y Matrix}
\usage{
gen_Y(PY_a, CLASS, N, J, P)
}
\arguments{
\item{PY_a}{A 3-dimensional cube containing probabilities for each class.}

\item{CLASS}{A numeric vector containing class labels for each respondent.}

\item{N}{Number of respondents.}

\item{J}{Number of response categories.}

\item{P}{Number of intervals for each response category.}
}
\value{
A matrix representing the generated categorical responses.
}
\description{
Generate a matrix of categorical responses based on probability cubes and class labels.

This function generates a matrix of categorical responses (\code{Y}) based on the probability cubes (\code{PY_a}) and class labels (\code{CLASS}).
}
