% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenerateAtable}
\alias{GenerateAtable}
\title{Generate Attribute Table}
\usage{
GenerateAtable(nClass, K, M)
}
\arguments{
\item{nClass}{Number of classes.}

\item{K}{Number of attributes.}

\item{M}{Levels of attribute.}
}
\value{
A list containing the generated tables.
}
\description{
Generate attribute-related tables based on specified parameters.

This function generates attribute-related tables including the attribute table (\code{Atable}), lower bound table (\code{LBtable}),
index table (\code{finalcols}), D-to-Q table (\code{DtoQtable}), adjacency table (\code{adjtable}), and item index table (\code{itemindexnum})
based on the provided parameters including the number of classes (\code{nClass}), the number of attributes (\code{K}),
the number of levels of attribute (\code{M}).
}
