% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getmode}
\alias{getmode}
\title{Get Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
The mode of the input vector as an integer.
}
\description{
Get the mode of a vector.

This function calculates the mode of a numeric vector.
}
