% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{identify_strictcheck}
\alias{identify_strictcheck}
\title{Check Strict Identifiability Condition}
\usage{
identify_strictcheck(DELTA, completematrix, M, K, ncomp, mainorder)
}
\arguments{
\item{DELTA}{Matrix representing the Delta matrix.}

\item{completematrix}{Matrix used to check block diagonal matrix.}

\item{M}{Levels of attribute.}

\item{K}{Number of attribute.}

\item{ncomp}{Number of parameters.}

\item{mainorder}{A vector representing the main order of the parameters.}
}
\value{
A boolean indicating whether the strict identifiability condition is satisfied.
}
\description{
Check if the Delta matrix satisfies the strict identifiability condition.

This function checks if the provided Delta matrix satisfies the strict identifiability condition based on the given parameters.
}
