% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_gen_bijectionvector}
\alias{inv_gen_bijectionvector}
\title{Inverse Generate Bijection Vector}
\usage{
inv_gen_bijectionvector(K, M, CL)
}
\arguments{
\item{K}{Number of attributes.}

\item{M}{Level of attribute.}

\item{CL}{Cumulative value.}
}
\value{
A numeric vector representing the inverse bijection vector.
}
\description{
Inverse function of generating the bijection vector for a given number of attributes and level of attribute.

This function calculates the inverse of the function that generates the bijection vector for a specified number of attributes (K),
level of attribute (M), and cumulative value (CL).
}
