% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m2LL}
\alias{m2LL}
\title{Compute the -2*log-likelihood for responses.}
\usage{
m2LL(N, J, P, nClass, Y, pis, PY_a)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{P}{Number of response categories.}

\item{nClass}{Number of latent classes.}

\item{Y}{A matrix of observations with respondents by items.}

\item{pis}{Vector of latent class proportion parameters.}

\item{PY_a}{Cube of the joint probability distribution.}
}
\value{
-2*log-likelihood.
}
\description{
This function computes the -2*log-likelihood for responses.
}
