% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myrnorm}
\alias{myrnorm}
\title{Generate Random Normal Samples}
\usage{
myrnorm(mu, sigma)
}
\arguments{
\item{mu}{A numeric vector specifying the mean of the normal distribution.}

\item{sigma}{A numeric vector specifying the standard deviation of the normal distribution.}
}
\value{
A numeric vector of random samples.
}
\description{
Generate random samples from a normal distribution with given mean and standard deviation.

This function generates random samples from a normal distribution with given mean (\code{mu}) and standard deviation (\code{sigma}).
}
