% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permuteAtableIndices}
\alias{permuteAtableIndices}
\title{Permute Atable Indices}
\usage{
permuteAtableIndices(nClass, K, M, vv, perm)
}
\arguments{
\item{nClass}{Number of classes.}

\item{K}{Number of attributes.}

\item{M}{Levels of attribute.}

\item{vv}{Vector of values.}

\item{perm}{Permutation vector.}
}
\value{
A vector representing the permuted indices of the attribute profile table.
}
\description{
Permute indices of the attribute profile table based on specified parameters.

This function permutes the indices of the attribute profile table (\code{Atable}) based on the provided parameters including
the number of classes (\code{nClass}), the number of attributes (\code{K}), the number of levels of attribute (\code{M}),
the vector of values (\code{vv}) and the permutation vector (\code{perm}).
}
