% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_D}
\alias{random_D}
\title{Generate Random Delta Matrix}
\usage{
random_D(J, K, w)
}
\arguments{
\item{J}{Number of items.}

\item{K}{Number of attributes.}

\item{w}{Probability parameter specifying sparsity.}
}
\value{
A random Delta matrix.
}
\description{
Generate a random Delta matrix with specified dimensions and parameters.

This function generates a random Delta matrix with specified dimensions and parameters,
satisfying identifiability constraints.
}
\examples{
J <- 10
K <- 3
w <- 0.5
}
