% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_D_strict}
\alias{random_D_strict}
\title{Generate Random Delta Matrix with Strict Identifiability}
\usage{
random_D_strict(J, K, M, t, DtoQtable)
}
\arguments{
\item{J}{Number of items.}

\item{K}{Number of attributes.}

\item{M}{Number of levels per attribute.}

\item{t}{Probability parameter specifying sparsity.}

\item{DtoQtable}{Table indicating the number of attributes per item.}
}
\value{
A random Delta matrix satisfying the strict identifiability condition.
}
\description{
Generate a random Delta matrix with strict identifiability condition and specified dimensions and parameters.

This function generates a random Delta matrix with strict identifiability condition.
}
