% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbernoulli_one}
\alias{rbernoulli_one}
\title{Generate Bernoulli Random Sample}
\usage{
rbernoulli_one(prob)
}
\arguments{
\item{prob}{A numeric value specifying the probability of success.}
}
\value{
An integer indicating the outcome of the Bernoulli trial (0 or 1).
}
\description{
Generate a Bernoulli random sample with given probability of success.

This function generates a single Bernoulli random sample with the specified probability (\code{prob}) of success.
}
