% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Generate Dirichlet Sample}
\usage{
rDirichlet(deltas)
}
\arguments{
\item{deltas}{A numeric vector specifying the parameters of the Dirichlet distribution.}
}
\value{
A numeric vector representing the generated Dirichlet sample.
}
\description{
Generate a random sample from a Dirichlet distribution with given parameters.

This function generates a single random sample from a Dirichlet distribution with parameters specified by the vector \code{deltas}.
}
