% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinomial}
\alias{rmultinomial}
\title{Generate Multinomial Sample}
\usage{
rmultinomial(ps)
}
\arguments{
\item{ps}{A numeric vector specifying the probabilities of each category.}
}
\value{
An unsigned integer representing the generated multinomial sample.
}
\description{
Generate a multinomial sample with given probabilities.

This function generates a single multinomial sample with the specified probabilities (\code{ps}).
}
