% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RTLL}
\alias{RTLL}
\title{Compute the log-likelihood ratio test statistic for response time.}
\usage{
RTLL(N, J, nClass, logt, Eta, Tau, Atable, H, sigma2_t, pis)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{nClass}{Number of latent classes.}

\item{logt}{A matrix of log-transformed response time.}

\item{Eta}{A vector of item response time parameters.}

\item{Tau}{A vector of latent speed coefficients.}

\item{Atable}{A matrix of item attribute profiles.}

\item{H}{A matrix of item-attribute response time coefficients.}

\item{sigma2_t}{A vector of item response time variances.}

\item{pis}{Vector of latent class proportion parameters.}
}
\value{
Log-likelihood ratio test statistic.
}
\description{
This function computes the log-likelihood ratio test statistic.
}
