% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rTruncNorm}
\alias{rTruncNorm}
\title{Generate Truncated Normal Sample}
\usage{
rTruncNorm(mean, sd, w, ps)
}
\arguments{
\item{mean}{A numeric value specifying the mean of the normal distribution.}

\item{sd}{A numeric value specifying the standard deviation of the normal distribution.}

\item{w}{A numeric value specifying the response category.}

\item{ps}{A numeric vector of cumulative probabilities.}
}
\value{
A numeric value representing the truncated normal sample.
}
\description{
Generate a truncated normal sample with specified mean, standard deviation, multinomial result, and cumulative probabilities
given the multinomial distribution.

This function generates a single truncated normal sample with the specified mean (\code{mean}), standard deviation (\code{sd}),
response category (\code{w}), and cumulative probabilities (\code{ps}).
}
