% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rTruncNorm_1b}
\alias{rTruncNorm_1b}
\title{Generate Truncated Normal Sample}
\usage{
rTruncNorm_1b(mu, sigma, y, c)
}
\arguments{
\item{mu}{A numeric value specifying the mean of the normal distribution.}

\item{sigma}{A numeric value specifying the standard deviation of the normal distribution.}

\item{y}{A numeric value specifying the truncation direction (0 for upper bound, 1 for lower bound).}

\item{c}{A numeric value specifying the truncation bound.}
}
\value{
A numeric value representing the truncated normal sample.
}
\description{
Generate a truncated normal sample with specified mean, standard deviation, truncation direction, and truncation bound.

This function generates a single truncated normal sample with the specified mean (\code{mu}), standard deviation (\code{sigma}),
truncation direction (\code{y}), and truncation bound (\code{c}).
}
