% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simY}
\alias{simY}
\title{Simulate Y Matrix}
\usage{
simY(N, J, M, nClass, CLASS, Atable, BETA, Kappa)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{M}{Levels of attribute.}

\item{nClass}{Number of classes.}

\item{CLASS}{A numeric vector containing class labels for each respondent.}

\item{Atable}{A matrix representing the attribute profile table.}

\item{BETA}{A matrix representing the item response coefficient matrix.}

\item{Kappa}{A matrix representing the item response threshold matrix.}
}
\value{
A matrix of (\code{N}) by (\code{J}) representing the simulated categorical responses.
}
\description{
Simulate a matrix of categorical responses based on provided parameters.

This function simulates a matrix of categorical responses (\code{Y}) based on the provided parameters including the number of respondents (\code{N}),
the number of items (\code{J}), the number of levels of attribute (\code{M}), the number of classes (\code{nClass}), class labels (\code{CLASS}),
attribute profile table (\code{Atable}), item response coefficient matrix (\code{BETA}), and item response threshold matrix (\code{Kappa}).
}
