% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simYast}
\alias{simYast}
\title{Simulate Y star.}
\usage{
simYast(N, J, Y, CLASS, PY_a, ABETA)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{Y}{Matrix of categorical responses.}

\item{CLASS}{Vector of class labels.}

\item{PY_a}{Cube of conditional probabilities.}

\item{ABETA}{Matrix of adjusted item response coefficients.}
}
\value{
A matrix of simulated categorical responses.
}
\description{
Simulate a matrix of categorical responses.

This function simulates a matrix of categorical responses using the provided parameters and an adjusted sampling technique.
}
