% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLCMRT.R
\name{slcmrt_Gibbs}
\alias{slcmrt_Gibbs}
\title{SLCMRT_Gibbs}
\usage{
slcmrt_Gibbs(Y, logt, K, M, burnin, chain_length, rep_each)
}
\arguments{
\item{Y}{A matrix of observations with respondents by items.}

\item{logt}{A matrix of log-transformed response times with respondents by items.}

\item{K}{An integer specifying the number of attributes.}

\item{M}{An integer specifying the number of levels of attributes.}

\item{burnin}{An integer specifying the number of burn-in iterations.}

\item{chain_length}{An integer specifying the total number of iterations.}

\item{rep_each}{An integer specifying the number of repetitions.}
}
\value{
A list containing estimated parameters including:
\itemize{
  \item est_Eta: Estimated item response time coefficients.
  \item est_Tau: Estimated person speed coefficients.
  \item est_CL: Estimated latent classes.
  \item est_pis: Estimated class proportions.
  \item est_H: Estimated item-attribute response time coefficient matrix.
  \item sum_D: Sum of D matrices over iterations.
  \item est_sigma2_eta: Estimated hyperparameter of variance.
  \item est_sigma2_t: Estimated hyperparameters of variance.
  \item est_D: Estimated item-attribute structure matrix.
  \item est_Beta: Estimated item response coefficients.
  \item est_ABETA: Estimated Alpha times Beta coefficient matrix.
}
}
\description{
Run the Gibbs sampling algorithm for fitting the SLCMRT model.
}
