% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_alpha}
\alias{Update_alpha}
\title{Update_alpha}
\usage{
Update_alpha(N, J, nClass, Y, CLASS, PY_a, pis, Atable, alpha, d0)
}
\arguments{
\item{N}{An integer specifying the number of observations.}

\item{J}{An integer specifying the number of items.}

\item{nClass}{An integer specifying the number of latent classes.}

\item{Y}{Matrix of categorical responses.}

\item{CLASS}{Vector of class labels.}

\item{PY_a}{A cube containing the conditional probabilities for each observation-feature combination and class.}

\item{pis}{Vector of latent class proportion parameters.}

\item{Atable}{Matrix representing the attribute profile table.}

\item{alpha}{Matrix of attribute profiles by respondents.}

\item{d0}{A vector of hyper parameters.}
}
\value{
A list containing the updated parameters:
\itemize{
  \item pis: Updated class proportions.
  \item CLASS: Updated class assignments.
  \item alpha: Updated alpha parameters.
}
}
\description{
Update the alpha parameters in the SLCM model.
}
