% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_DB}
\alias{Update_DB}
\title{Update_DB}
\usage{
Update_DB(
  J,
  M,
  K,
  alpha,
  nClass,
  Yast,
  BETA,
  DELTA,
  ABETA,
  Atable,
  sigma_B,
  omega
)
}
\arguments{
\item{J}{An integer specifying the number of rows.}

\item{M}{An integer specifying the number of components.}

\item{K}{An integer specifying the number of factors.}

\item{alpha}{A matrix containing alpha values.}

\item{nClass}{An integer specifying the number of classes.}

\item{Yast}{A matrix containing Yast values.}

\item{BETA}{A matrix containing BETA values.}

\item{DELTA}{A matrix containing DELTA values.}

\item{ABETA}{A matrix containing ABETA values.}

\item{Atable}{A matrix containing Atable values.}

\item{sigma_B}{A double specifying the parameter sigma_B.}

\item{omega}{A double specifying the parameter omega.}
}
\value{
A list containing updated BETA, DELTA, and ABETA values.
}
\description{
Update BETA and DELTA parameters.
}
