% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_Eta}
\alias{Update_Eta}
\title{Update Eta}
\usage{
Update_Eta(N, J, Tau, logt, alpha, H, sigma2_t, sigma2_eta)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{Tau}{A numeric vector of latent speed coefficients.}

\item{logt}{A numeric matrix of log-transformed response time.}

\item{alpha}{A numeric matrix of attribute profiles by respondents.}

\item{H}{A numeric matrix of item-attribute response time coefficients.}

\item{sigma2_t}{A numeric vector of item response time variances.}

\item{sigma2_eta}{Variance parameter for the Eta sampling.}
}
\value{
A numeric vector of updated Eta values.
}
\description{
Update Eta using MCMC sampling.

This function updates Eta using MCMC sampling.
}
