% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_HDB_strict}
\alias{Update_HDB_strict}
\title{Update H, D and B matrix}
\usage{
Update_HDB_strict(
  J,
  N,
  M,
  K,
  nClass,
  sigma2_h,
  alpha,
  sigma2_t,
  logt,
  Atable,
  H,
  Eta,
  Tau,
  DELTA,
  omega,
  Yast,
  BETA,
  ABETA,
  sigma_B,
  main,
  mainorder,
  completematrix,
  LBtable,
  Bindices
)
}
\arguments{
\item{J}{Number of items.}

\item{N}{Number of respondents.}

\item{M}{Levels of attribute.}

\item{K}{Number of attributes.}

\item{nClass}{Number of classes.}

\item{sigma2_h}{Variance parameter for H.}

\item{alpha}{Matrix of attribute profiles by respondents.}

\item{sigma2_t}{Vector of variance parameters for the response time.}

\item{logt}{Matrix of log-transformed response time.}

\item{Atable}{Matrix representing the attribute profile table.}

\item{H}{Matrix of item-attribute response time coefficients.}

\item{Eta}{Vector of item response time coefficients.}

\item{Tau}{Vector of latent speed coefficients.}

\item{DELTA}{Matrix representing the item-attribute structure.}

\item{omega}{Hyperparameter.}

\item{Yast}{Matrix representing simulated responses.}

\item{BETA}{Matrix representing the item response coefficient matrix.}

\item{ABETA}{Matrix representing the item response coefficient matrix.}

\item{sigma_B}{Variance parameter for B.}

\item{main}{Vector representing the column index of the main-effect parameters.}

\item{mainorder}{Vector representing the index of the beta coefficients.}

\item{completematrix}{Matrix used to check the identifiablity constraints.}

\item{LBtable}{Matrix representing the lower bound table.}

\item{Bindices}{Vector representing the latent class indices.}
}
\value{
None. The function updates the HDB matrix in place.
}
\description{
Update the HDB matrix with strict identifiability condition.

This function updates the HDB matrix with strict identifiability condition using the provided parameters.
The H coefficients are negative in default.
}
