% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_sigma2_t}
\alias{Update_sigma2_t}
\title{Update Sigma2_t}
\usage{
Update_sigma2_t(N, J, Eta, Tau, logt, alpha, H, a_t, b_t)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{Eta}{A numeric vector of item response time parameters.}

\item{Tau}{A numeric vector of the latent speed Tau values.}

\item{logt}{A numeric matrix of log-transformed response time.}

\item{alpha}{A numeric matrix of attribute profiles by respondents.}

\item{H}{A numeric matrix of item-attribute response time coefficients.}

\item{a_t}{Shape parameter for the gamma distribution.}

\item{b_t}{Rate parameter for the gamma distribution.}
}
\value{
A numeric vector of updated Sigma2_t values.
}
\description{
Update Sigma2_t using MCMC sampling.

This function updates Sigma2_t using MCMC sampling.
}
