% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Update_Tau}
\alias{Update_Tau}
\title{Update the latent speed coefficients Tau}
\usage{
Update_Tau(N, J, Eta, sigma2_t, logt, alpha, H)
}
\arguments{
\item{N}{Number of respondents.}

\item{J}{Number of items.}

\item{Eta}{A numeric vector of item response time parameters.}

\item{sigma2_t}{A numeric vector of item response time variances.}

\item{logt}{A numeric matrix of log-transformed response time.}

\item{alpha}{A numeric matrix of attribute profiles by respondents.}

\item{H}{A numeric matrix of item-attribute response time coefficients.}
}
\value{
A numeric vector.
}
\description{
Update the latent speed coefficients Tau using MCMC sampling.

This function updates the latent speed coefficients Tau using MCMC sampling.
}
